requirements.txt
aio-pika==1.5.1
aiohttp==2.3.1
astroid==1.5.3
async-timeout==2.0.0
certifi==2016.2.28
chardet==3.0.4
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
msgpack-python==0.4.8
multidict==3.3.0
numpy==1.13.3
pandas==0.20.3
pika==0.10.0
py==1.4.34
pycodestyle==2.3.1
pydocstyle==2.1.1
pylint==1.7.4
pytest==3.2.3
pytest-aiohttp==0.1.3
python-dateutil==2.6.1
pytz==2017.2
shortuuid==0.5.0
six==1.11.0
snowballstemmer==1.2.1
wrapt==1.10.11
yarl==0.13.0
